commit all

+++ mmdet/models/anchor_heads/rpn_head.py
@@ -142,7 +142,7 @@ class RPNHead(AnchorHead):
         KLD = KLD.mean()
         return KLD
     def _reparameterization(self, input):
-        half_size = input.size(1)/2
+        half_size = int(input.size(1)/2)
         print(half_size)
         mu = input[:,:half_size]
         if self.train:
